Pull install size out of fetch_formulae
