+++ internal/origins/drivers/brew/install_size.go
@@ -1,0 +1,29 @@
+package brew
+
+import (
+	"io/fs"
+	"path/filepath"
+)
+
+func getInstallSize(dir string) (int64, error) {
+	var total int64
+
+	err := filepath.WalkDir(dir, func(_ string, dir fs.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+
+		if !dir.IsDir() {
+			info, err := dir.Info()
+			if err != nil {
+				return err
+			}
+
+			total += info.Size()
+		}
+
+		return nil
+	})
+
+	return total, err
+}
